+++ cliffordcorner.yml
@@ -1,0 +1,73 @@
+album:
+  type: pool
+  id: 31386902@N00
+  title: "Photographs"
+  description: "Photographs by, about, and related to the Clifford Family."
+  sort_order: 1
+album:
+  type: album
+  id: 72157645299291057
+  title: "Sabbatical 2014"
+  description: "Photographs from our family trips to Mexico and Michigan."
+  sort_order: 2
+album:
+  type: album
+  id: 72157624473398430
+  title: "Sabbatical 2010"
+  description: "Photographs from our adventures through Turkey, Greece and Croatia."
+  sort_order: 3
+album:
+  type: album
+  id: 72157621459686057
+  title: "Wedding"
+  description: "Photographs and notes from our Wedding."
+  sort_order: 4
+album:
+  type: album
+  id: 72157620904653603
+  title: "Honeymoon"
+  description: "Our relaxing honeymoon adventure to Roatan, Honduras."
+  sort_order: 5
+album:
+  type: album
+  id: 72157606701085631
+  title: "Peru"
+  description: "A trek around Salkantey, ending at Machu Picchu."
+  sort_order: 6
+album:
+  type: album
+  id: 72157602427960981
+  title: "Nepal"
+  description: "A trek through the Everest Region in Nepal. Including stops in Bangkok and Kathmandu."
+  sort_order: 7
+album:
+  type: album
+  id: 72157594373088876
+  title: "Australia"
+  description: "Travels through Australia and Tasmania."
+  sort_order: 8
+album:
+  type: album
+  id: 1157473
+  title: "Bhutan"
+  description: "Exploring the Himilayan Kingdom of Bhutan, Angkor Wat, and Bangkok."
+  sort_order: 9
+album:
+  type: album
+  id: 707782
+  title: "Chile"
+  description: "Quick adventure down to the Thin Country (Chile)."
+  sort_order: 10
+album:
+  type: microsite
+  id: africa
+  uri: /tavellog/africa
+  title: "Africa"
+  description: "African Safari Adventure."
+  sort_order: 11
+album:
+  type: album
+  id: 72157594259055081
+  title: "Moab"
+  description: "Hiking, Mountain Biking, and otherwise adventuring solo around Utah."
+  sort_order: 12
