--- photograph_album_tag.rb
@@ -1,19 +1,0 @@
-class PhotographAlbumTag
-	include Comparable	
-
-	attr_accessor :tag
-	attr_accessor :count
-
-	def initialize(tag = nil)
-		@tag = tag
-		@count = 1
-	end
-
-	def <=>(other)
-		if other.count == count
-			tag <=> other.tag
-		else
-			other.count <=> count
-		end
-	end
-end
