--- photograph_size.rb
@@ -1,8 +1,0 @@
-module PhotographSize
-	UNDEFINED = -1
-	THUMBNAIL = 0
-	SMALL = 1
-	MEDIUM = 2
-	LARGE = 3
-	ORIGINAL = 4
-end
